+++ CHANGELOG.md
@@ -9,3 +9,12 @@
 
 ### 0.0.3
 - Added schema file to package
+
+### 0.0.4
+- Points are not required for a valid OSW dataset
+
+### 0.0.5
+- Support for multi-level geojson file
+- Now handles the following two folder structures when unzipped abc.zip
+  1. abc\{nodes, edges, points}.geojson
+  2. {nodes, edges, points}.geojson